backend/requirements.txt
# CLI agent + API server
langchain
langchain-openai
supabase
python-dotenv
rapidfuzz
fastapi
uvicorn[standard]
